+++ src/App.js
@@ -1,21 +1,8 @@
-import React, { Suspense, useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { saveUser } from "./redux/user/actions";
+import React, { Suspense } from "react";
 import Header from "./pages/Header";
 import AppRouter from "./routes/AppRouter";
 
 function App() {
-  const dispatch = useDispatch();
-  const { error } = useSelector((state) => state);
-
-  useEffect(() => {
-    dispatch(saveUser());
-  }, []);
-
-  useEffect(() => {
-    if (error) error.map((err) => window.alert(err));
-  }, [error]);
-
   return (
     <Suspense fallback={<div>Loading...</div>}>
       <Header />