+++ src/tests/pages/components/Logo.test.js
@@ -1,0 +1,9 @@
+import { shallow } from "enzyme";
+import Logo from "../../../pages/components/Logo";
+
+describe("<Logo />", () => {
+  test("match snapshot", () => {
+    const wrapper = shallow(<Logo />);
+    expect(wrapper).toMatchSnapshot();
+  });
+});